requirements.txt
pydicom
pyyaml
pika
pandas
dicompyler-core
requests